+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoRuntime.java
@@ -340,7 +340,8 @@ public final class GeckoRuntime implements Parcelable {
         final GeckoBundle bundle = new GeckoBundle(3);
         bundle.putString("locationUri", webExtension.location);
         bundle.putString("id", webExtension.id);
-        bundle.putBoolean("allowContentMessaging", webExtension.allowContentMessaging);
+        bundle.putBoolean("allowContentMessaging",
+                (webExtension.flags & WebExtension.Flags.ALLOW_CONTENT_MESSAGING) > 0);
 
         mWebExtensionDispatcher.registerWebExtension(webExtension);