+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/WebExtensionTest.kt
@@ -110,28 +110,18 @@ class WebExtensionTest : BaseSessionTest() {
             }
         }
 
-        val messaging = createWebExtension(background, messageDelegate)
+        val messaging = WebExtension(if (background) MESSAGING_BACKGROUND else MESSAGING_CONTENT,
+                "{${UUID.randomUUID()}}", !background)
+        if (background) {
+            messaging.setMessageDelegate(messageDelegate, "browser")
+        } else {
+            sessionRule.session.setMessageDelegate(messageDelegate, "browser");
+        }
+
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(messaging))
         sessionRule.waitForResult(messageResult)
 
         sessionRule.waitForResult(sessionRule.runtime.unregisterWebExtension(messaging))
-    }
-
-    private fun createWebExtension(background: Boolean,
-                                   messageDelegate: WebExtension.MessageDelegate): WebExtension {
-        val webExtension: WebExtension
-        val uuid = "{${UUID.randomUUID()}}"
-
-        if (background) {
-            webExtension = WebExtension(MESSAGING_BACKGROUND, uuid, WebExtension.Flags.NONE)
-            webExtension.setMessageDelegate(messageDelegate, "browser")
-        } else {
-            webExtension = WebExtension(MESSAGING_CONTENT, uuid,
-                    WebExtension.Flags.ALLOW_CONTENT_MESSAGING)
-            sessionRule.session.setMessageDelegate(messageDelegate, "browser");
-        }
-
-        return webExtension
     }
 
     @Test
@@ -202,7 +192,13 @@ class WebExtensionTest : BaseSessionTest() {
             }
         }
 
-        val messaging = createWebExtension(background, messageDelegate)
+        val messaging = WebExtension(if (background) MESSAGING_BACKGROUND else MESSAGING_CONTENT,
+                "{${UUID.randomUUID()}}", !background)
+        if (background) {
+            messaging.setMessageDelegate(messageDelegate, "browser")
+        } else {
+            sessionRule.session.setMessageDelegate(messageDelegate, "browser");
+        }
 
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(messaging))
         sessionRule.waitForResult(result)
@@ -284,7 +280,13 @@ class WebExtensionTest : BaseSessionTest() {
             }
         }
 
-        messaging = createWebExtension(background, messageDelegate)
+        messaging = WebExtension(if (background) MESSAGING_BACKGROUND else MESSAGING_CONTENT,
+                "{${UUID.randomUUID()}}", !background)
+        if (background) {
+            messaging.setMessageDelegate(messageDelegate, "browser")
+        } else {
+            sessionRule.session.setMessageDelegate(messageDelegate, "browser");
+        }
 
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(messaging))
         sessionRule.waitForResult(result)
@@ -364,7 +366,13 @@ class WebExtensionTest : BaseSessionTest() {
             }
         }
 
-        messaging = createWebExtension(background, messageDelegate)
+        messaging = WebExtension(if (background) MESSAGING_BACKGROUND else MESSAGING_CONTENT,
+                "{${UUID.randomUUID()}}", !background)
+        if (background) {
+            messaging.setMessageDelegate(messageDelegate, "browser")
+        } else {
+            sessionRule.session.setMessageDelegate(messageDelegate, "browser");
+        }
 
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(messaging))
         sessionRule.waitForResult(result)
@@ -433,7 +441,7 @@ class WebExtensionTest : BaseSessionTest() {
         }
 
         messaging = WebExtension("resource://android/assets/web_extensions/messaging-iframe/",
-                "{${UUID.randomUUID()}}", WebExtension.Flags.ALLOW_CONTENT_MESSAGING)
+                "{${UUID.randomUUID()}}", true)
         sessionRule.session.setMessageDelegate(messageDelegate, "browser");
 
         sessionRule.waitForResult(sessionRule.runtime.registerWebExtension(messaging))