+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -304,4 +304,4 @@ exclude: true
 [65.24]: ../CrashReporter.html#sendCrashReport-android.content.Context-android.os.Bundle-java.lang.String-
 [65.25]: ../GeckoResult.html
 
-[api-version]: 3b910c2e6b3df9bd9926cf4e54b37a5a08bdb64a
+[api-version]: d51e6bd6f7fe004cbe5856e892ec538e2404c3fc